requirements-server.txt
boto3
fastapi[standard]==0.115.5
pillow==10.4.0
numpy==1.26.4